@@ -36,46 +38,57 @@
   "homepage": "https://github.com/bormgruppe/ng2-pdf-viewer#readme",
   "peerDependencies": {
     "@angular/core": ">=5.0.0",
-    "pdfjs-dist": "2.2.228",
-    "@types/pdfjs-dist": "^0.1.2"
+    "pdfjs-dist": "^2.2.228",
+    "@types/pdfjs-dist": "^2.1.0"
   },
   "dependencies": {
-    "@types/pdfjs-dist": "^0.1.2",
-    "pdfjs-dist": "2.2.228"
+    "@types/pdfjs-dist": "^2.1.0",
+    "pdfjs-dist": "^2.2.228"
   },
   "devDependencies": {
-    "@angular/animations": "^5.2.0",
-    "@angular/cdk": "^5.1.1",
-    "@angular/cli": "1.6.6",
-    "@angular/common": "^5.2.0",
-    "@angular/compiler": "^5.2.0",
-    "@angular/compiler-cli": "^5.2.0",
-    "@angular/core": "^5.2.0",
-    "@angular/forms": "^5.2.0",
-    "@angular/http": "^5.2.0",
-    "@angular/language-service": "^5.2.0",
-    "@angular/material": "^5.1.1",
-    "@angular/platform-browser": "^5.2.0",
-    "@angular/platform-browser-dynamic": "^5.2.0",
-    "@angular/router": "^5.2.0",
-    "@types/jasmine": "~2.8.3",
-    "@types/jasminewd2": "~2.0.2",
-    "@types/node": "^6.0.96",
-    "codelyzer": "^4.0.1",
-    "core-js": "^2.4.1",
-    "jasmine-core": "~2.8.0",
+    "@angular-devkit/build-angular": "~0.803.6",
+    "@angular/animations": "^8.0.0",
+    "@angular/cdk": "^8.2.1",
+    "@angular/cli": "~8.3.6",
+    "@angular/common": "~8.2.8",
+    "@angular/compiler": "~8.2.8",
+    "@angular/compiler-cli": "~8.2.8",
+    "@angular/core": "~8.2.8",
+    "@angular/forms": "~8.2.8",
+    "@angular/language-service": "~8.2.8",
+    "@angular/material": "^8.2.1",
+    "@angular/platform-browser": "~8.2.8",
+    "@angular/platform-browser-dynamic": "~8.2.8",
+    "@angular/router": "~8.2.8",
+    "@types/jasmine": "~3.4.0",
+    "@types/jasminewd2": "~2.0.3",
+    "@types/node": "^12.7.7",
+    "codelyzer": "^5.0.0",
+    "jasmine-core": "~3.5.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~2.0.0",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "ng-packagr": "^1.6.0",
-    "protractor": "~5.1.2",
-    "rxjs": "^5.5.6",
-    "ts-node": "~4.1.0",
-    "tslint": "~5.9.1",
-    "typescript": "~2.6.2",
-    "zone.js": "^0.8.19"
+    "karma": "~4.4.0",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "~2.1.0",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.0",
+    "ng-packagr": "^5.2.0",
+    "protractor": "~5.4.0",
+    "rxjs": "~6.5.3",
+    "ts-node": "~8.4.1",
+    "tslib": "^1.9.0",
+    "tslint": "~5.20.0",
+    "typescript": "~3.5.3",
+    "webpack-bundle-analyzer": "^3.3.2",
+    "zone.js": "~0.10.2"
+  },
+  "$schema": "./node_modules/ng-packagr/package.schema.json",
+  "ngPackage": {
+    "lib": {
+      "entryFile": "public_api.ts"
+    },
+    "whitelistedNonPeerDependencies": [
+      "@types/pdfjs-dist",
+      "pdfjs-dist"
+    ]
   }
 }
